seperating the main execution from the functions
